portal/api/requirements.txt
azure-functions
bonnette
fastapi
azure-identity
azure-keyvault-secrets
azure-cosmos